+++ gradle.properties
@@ -24,3 +24,4 @@ org.gradle.jvmargs=-Xmx2G
 # Development QOL
 	modmenu_version=3.2.5
 	rei_version=8.3.588
+
